remove analysis/stitch.py
